FourMomentumBase

Documentation for FourMomentumBase.

FourMomentumBase.EXYZ

FourMomentumBase.azimuthal_angle

FourMomentumBase.boost_beta

FourMomentumBase.boost_gamma

FourMomentumBase.coordinate_names

FourMomentumBase.coordinate_system

FourMomentumBase.cos_phi

FourMomentumBase.cos_theta

FourMomentumBase.energy

FourMomentumBase.invariant_mass

FourMomentumBase.invariant_mass2

FourMomentumBase.magnitude

FourMomentumBase.magnitude2

FourMomentumBase.minus_component

FourMomentumBase.plus_component

FourMomentumBase.polar_angle

FourMomentumBase.pseudorapidity

FourMomentumBase.px

FourMomentumBase.py

FourMomentumBase.pz

FourMomentumBase.rapidity

FourMomentumBase.sin_phi

FourMomentumBase.transverse_mass

FourMomentumBase.transverse_mass2

FourMomentumBase.transverse_momentum

FourMomentumBase.transverse_momentum2

FourMomentumBase.EXYZ — Type

EXYZ <: AbstractCoordinateSystem

Cartesian coordinate system for four-momenta. Using this requires the implementation of the following interface functions:

px(::CustomFourMomentum)
py(::CustomFourMomentum)
pz(::CustomFourMomentum)
energy(::CustomFourMomentum)

FourMomentumBase.azimuthal_angle — Method

azimuthal_angle(::EXYZ, mom)

Return the phi angle for a given four-momentum, i.e. the azimuthal angle of its spatial components in spherical coordinates.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to atan(py,px).

Note

The spherical coordinates are defined w.r.t. to the 3-axis.

FourMomentumBase.boost_beta — Method

boost_beta(::EXYZ, mom )

Return magnitude of the beta vector for a given four-momentum, i.e. the magnitude of the four-momentum divided by its 0-component.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to sqrt(px^2 + py^2 + pz^2)/E.

FourMomentumBase.boost_gamma — Method

boost_gamma(::EXYZ,mom)

Return the relativistic gamma factor for a given four-momentum, i.e. the inverse square root of one minus the beta vector squared.

Example

If (E,px,py,pz) is a four-momentum with beta vector β, this is equivalent to 1/sqrt(1- β^2).

FourMomentumBase.coordinate_names — Function

coordinate_names(::AbstractCoordinateSystem)::Tuple{Symbols}

FourMomentumBase.coordinate_system — Function

coordinate_system(::MomType)::CS where {CS <: AbstractCoordinateSystem}

FourMomentumBase.cos_phi — Method

cos_phi(::EXYZ, mom)

Return the cosine of the azimuthal angle for a given four-momentum.

Note

This is an equivalent but faster version of cos(azimuthal_angle(mom)); see azimuthal_angle.

FourMomentumBase.cos_theta — Method

cos_theta(::EXYZ, mom)

Return the cosine of the theta angle for a given four-momentum.

Note

This is an equivalent but faster version of cos(polar_angle(::EXYZ, mom)); see polar_angle.

FourMomentumBase.energy — Function

energy(mom)

TBW

FourMomentumBase.invariant_mass — Method

invariant_mass(::EXYZ,mom)

Return the the invariant mass of a given four-momentum, i.e. the square root of the minkowski dot with itself.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to sqrt(E^2 - (px^2 + py^2 + pz^2)).

Note

If the squared invariant mass m2 is negative, -sqrt(-m2) is returned.

FourMomentumBase.invariant_mass2 — Method

invariant_mass2(::EXYZ,mom)

Return the squared invariant mass of a given four-momentum, i.e. the minkowski dot with itself.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to E^2 - (px^2 + py^2 + pz^2).

FourMomentumBase.magnitude — Method

magnitude(::EXYZ,mom)

Return the magnitude of a given four-momentum, i.e. the euklidian norm spatial components.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to sqrt(px^2 + py^2 + pz^2).

Warning

This function differs from a similar function for the TLorentzVector used in the famous ROOT library.

FourMomentumBase.magnitude2 — Method

magnitude2(::EXYZ, mom)

Return the square of the magnitude of a given four-momentum, i.e. the sum of the squared spatial components.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to px^2+ py^2 + pz^2.

Warning

This function differs from a similar function for the TLorentzVector used in the famous ROOT library.

FourMomentumBase.minus_component — Method

minus_component(::EXYZ, mom)

Return the minus component for a given four-momentum in light-cone coordinates.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to (E-pz)/2.

Note

The light-cone coordinates are defined w.r.t. to the 3-axis.

Warning

The definition `p^- := (E - p_z)/2 differs from the usual definition of light-cone coordinates in general relativity.

FourMomentumBase.plus_component — Method

plus_component(::EXYZ, mom)

Return the plus component for a given four-momentum in light-cone coordinates.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to (E+pz)/2.

Note

The light-cone coordinates are defined w.r.t. to the 3-axis.

Warning

The definition `p^+ := (E + p_z)/2 differs from the usual definition of light-cone coordinates in general relativity.

FourMomentumBase.polar_angle — Method

polar_angle(::EXYZ,mom)

Return the theta angle for a given four-momentum, i.e. the polar angle of its spatial components in spherical coordinates.

Example

If (E,px,py,pz) is a four-momentum with magnitude rho, this is equivalent to arccos(pz/rho), which is also equivalent to arctan(sqrt(px^2+py^2)/pz).

Note

The spherical coordinates are defined w.r.t. to the 3-axis.

FourMomentumBase.pseudorapidity — Method

pseudorapidity(::EXYZ, mom)

TBW

FourMomentumBase.px — Function

px(mom)

TBW

FourMomentumBase.py — Function

py(mom)

TBW

FourMomentumBase.pz — Function

pz(mom)

TBW

FourMomentumBase.rapidity — Method

rapidity(::EXYZ, mom)

Return the rapidity for a given four-momentum.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to 0.5*log((E+pz)/(E-pz)).

Note

The transverse components are defined w.r.t. to the 3-axis.

FourMomentumBase.sin_phi — Method

sin_phi(::EXYZ,mom)

Return the sine of the azimuthal angle for a given four-momentum.

Note

This is an equivalent but faster version of sin(azimuthal_angle(mom)); see azimuthal_angle.

FourMomentumBase.transverse_mass — Method

transverse_mass(::EXYZ,mom)

Return the transverse momentum for a given four-momentum, i.e. the square root of its squared transverse mass.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to sqrt(E^2 - pz^2).

Note

The transverse components are defined w.r.t. to the 3-axis.

Note

If the squared transverse mass mt2 is negative, -sqrt(-mt2) is returned.

FourMomentumBase.transverse_mass2 — Method

transverse_mass2(::EXYZ, mom)

Return the squared transverse mass for a given four-momentum, i.e. the difference of its squared 0- and 3-component.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to E^2 - pz^2.

Note

The transverse components are defined w.r.t. to the 3-axis.

FourMomentumBase.transverse_momentum — Method

transverse_momentum(::EXYZ, mom)

Return the transverse momentum for a given four-momentum, i.e. the magnitude of its transverse components.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to sqrt(px^2 + py^2).

Note

The transverse components are defined w.r.t. to the 3-axis.

FourMomentumBase.transverse_momentum2 — Method

transverse_momentum2(::EXYZ,mom)

Return the squared transverse momentum for a given four-momentum, i.e. the sum of its squared 1- and 2-component.

Example

If (E,px,py,pz) is a four-momentum, this is equivalent to px^2 + py^2.

Note

The transverse components are defined w.r.t. to the 3-axis.